+++ tests/testthat/test-model-fit.R
@@ -23,3 +23,9 @@ test_that("setting different rng_seed returns different output", {
     a_fit_sample_null$sample$artnum_t1_out[1]
   )
 })
+
+test_that("exceeding maximum iterations throws a warning", {
+  expect_warning(fit_tmb(a_tmb_inputs, outer_verbose = FALSE, max_iter = 5),
+                 "iteration limit reached")
+})
+