+++ NEWS.md
@@ -3,6 +3,8 @@
 * Add model input validation step to hintr_run_model().
 * Implement random number seed argument to sample_tmb() to return exact same results.
 * Silent fitting for hintr_run_model() unless option outer_verbose = TRUE or inner_verbose = TRUE.
+* Specify maximum number of iterations for fit_tmb().
+* Warning if fit_tmb() convergence error.
 
 # naomi 0.0.20